// src/interfaces/cadastroProfissionaisInterface.tsx
export interface CadastroProfissionaisInterface {
    id: number;
    nome: string;
    cpf: string;
    email: string;
    celular: string;
    dataNascimento: string;
    cidade: string;
    estado: string;
    pais: string;
    rua: string;
    numero: string;
    bairro: string;
    cep: string;
    complemento: string;
    senha: string;
    salario: string;
}